+++ AddOns/Syndicator/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Syndicator
 
-## [224](https://github.com/Baganator/Syndicator/tree/224) (2025-08-02)
-[Full Changelog](https://github.com/Baganator/Syndicator/compare/223...224) 
+## [225](https://github.com/Baganator/Syndicator/tree/225) (2025-08-07)
+[Full Changelog](https://github.com/Baganator/Syndicator/compare/224...225) 
 
-- Revert fallback on ATT xpac check, as the rationale for the change is missing  
+- Remove C\_PlayerInfo.HasAccountInventoryLock use (replaced with C\_Bank function)